+++ examples/showcase/CHANGELOG.md
@@ -1,5 +1,18 @@
 # showcase
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @meshql/core@0.6.0
+  - @meshql/access@0.2.4
+  - @meshql/client@0.5.2
+  - @meshql/http@0.5.2
+  - @meshql/integrity@0.5.2
+  - @meshql/sqlite@0.5.2
+  - @meshql/upload@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes